budget - store not value per day but value for semi-interval

+++ budget/budget.cpp
@@ -58,7 +58,7 @@ int main() {
 
     switch (ParseQueryType(query_type)) {
       case QueryType::COMPUTE_INCOME:
-        cout << budget_system.ComputeIncome(from, to) << endl;
+        cout << static_cast<double>(budget_system.ComputeIncome(from, to)) << endl;
         break;
       case QueryType::EARN:
         IncomeValue income;
